dnbind: replace tabs with spaces inside commands
tabs are treated as escape characters by DNOS, sending back the
autocomplete list and with it a new prompt '#' marking the end
of the list, that breaks internal command result waiting logic.

sending a new command while previous one is stil in progress can
lead to DNOS silently ignoring it.

a simple but efficient solution is replacing all tab characters
of input commands with spaces.

also this patch increases debug logs verbosity for ease of future
issues analysis.

Jobs: RS-1387

Signed-off-by: Marius Ionescu <mionescu@drivenets.com>